+++ lib/analyzers/typescript/types/IFileFilter.ts
@@ -1,1 +1,0 @@
-export type IFileFilter = (filename: string) => boolean;